Renamed projects and changed namespaces
