+++ CodeMonkeysMVVM.sln
@@ -48,11 +48,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Logging.Batching", "src\Log
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Logging.File", "src\Logging\Logging.File\Logging.File.csproj", "{3A77AD67-9356-40B3-909D-EC5699B9AEE4}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "DialogService", "DialogService", "{83518436-0476-4F4D-BE28-97E605BFF297}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Dialogs", "Dialogs", "{83518436-0476-4F4D-BE28-97E605BFF297}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DialogService", "src\DialogService\DialogService\DialogService.csproj", "{23A5E1A2-1506-4663-A68D-5FFD577B6999}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Dialogs", "src\Dialogs\Dialogs\Dialogs.csproj", "{37A8B3CF-C5F3-4B47-A497-8204E47AD1B9}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DialogService.Xamarin.Forms", "src\DialogService\DialogService.Xamarin.Forms\DialogService.Xamarin.Forms.csproj", "{67D3003B-4889-4204-B6A0-62E513B346F4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dialogs.Xamarin.Forms", "src\Dialogs\Dialogs.Xamarin.Forms\Dialogs.Xamarin.Forms.csproj", "{26F8340F-9164-4F6C-B9F0-091C25701BC7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -116,14 +116,14 @@ Global
 		{3A77AD67-9356-40B3-909D-EC5699B9AEE4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3A77AD67-9356-40B3-909D-EC5699B9AEE4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3A77AD67-9356-40B3-909D-EC5699B9AEE4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{23A5E1A2-1506-4663-A68D-5FFD577B6999}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{23A5E1A2-1506-4663-A68D-5FFD577B6999}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{23A5E1A2-1506-4663-A68D-5FFD577B6999}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{23A5E1A2-1506-4663-A68D-5FFD577B6999}.Release|Any CPU.Build.0 = Release|Any CPU
-		{67D3003B-4889-4204-B6A0-62E513B346F4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{67D3003B-4889-4204-B6A0-62E513B346F4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{67D3003B-4889-4204-B6A0-62E513B346F4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{67D3003B-4889-4204-B6A0-62E513B346F4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{37A8B3CF-C5F3-4B47-A497-8204E47AD1B9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{37A8B3CF-C5F3-4B47-A497-8204E47AD1B9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{37A8B3CF-C5F3-4B47-A497-8204E47AD1B9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{37A8B3CF-C5F3-4B47-A497-8204E47AD1B9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{26F8340F-9164-4F6C-B9F0-091C25701BC7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{26F8340F-9164-4F6C-B9F0-091C25701BC7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{26F8340F-9164-4F6C-B9F0-091C25701BC7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{26F8340F-9164-4F6C-B9F0-091C25701BC7}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -140,8 +140,8 @@ Global
 		{5B3AC20B-FA15-4985-BCB9-437C9162D3A8} = {A8EDB898-CC2D-4EF5-9CE5-CC48BDD51DE4}
 		{F0B5FE26-093B-4035-AE0D-CE561CF09B4A} = {A8EDB898-CC2D-4EF5-9CE5-CC48BDD51DE4}
 		{3A77AD67-9356-40B3-909D-EC5699B9AEE4} = {A8EDB898-CC2D-4EF5-9CE5-CC48BDD51DE4}
-		{23A5E1A2-1506-4663-A68D-5FFD577B6999} = {83518436-0476-4F4D-BE28-97E605BFF297}
-		{67D3003B-4889-4204-B6A0-62E513B346F4} = {83518436-0476-4F4D-BE28-97E605BFF297}
+		{37A8B3CF-C5F3-4B47-A497-8204E47AD1B9} = {83518436-0476-4F4D-BE28-97E605BFF297}
+		{26F8340F-9164-4F6C-B9F0-091C25701BC7} = {83518436-0476-4F4D-BE28-97E605BFF297}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {7EDC70E9-1A61-46BF-839A-6262FB9AC289}
