+++ src/Core/Dialogs/IDialogService.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 
-namespace CodeMonkeys.Core.DialogService
+namespace CodeMonkeys.Core.Dialogs
 {
     public interface IDialogService
     {
