+++ src/Dialogs/Dialogs.Xamarin.Forms/Dialogs.Xamarin.Forms.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>netstandard2.1</TargetFramework>
-    <RootNamespace>CodeMonkeys.DialogService.Xamarin.Forms</RootNamespace>
+    <RootNamespace>CodeMonkeysDialogs.Xamarin.Forms</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
@@ -11,7 +11,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\Core\Core.csproj" />
-    <ProjectReference Include="..\DialogService\DialogService.csproj" />
+    <ProjectReference Include="..\Dialogs\Dialogs.csproj" />
   </ItemGroup>
 
 </Project>
