+++ src/Dialogs/Dialogs/Dialogs.csproj
@@ -1,8 +1,8 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>netstandard2.1</TargetFramework>
-    <RootNamespace>CodeMonkeys.DialogService</RootNamespace>
+    <RootNamespace>CodeMonkeys.Dialogs</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
