+++ src/Dialogs/Dialogs/Models/AlertDialog.cs
@@ -1,4 +1,4 @@
-﻿namespace CodeMonkeys.DialogService
+﻿namespace CodeMonkeys.Dialogs
 {
     public class AlertDialog : Dialog
     {
