+++ src/Dialogs/Dialogs/Models/ConfirmationDialog.cs
@@ -1,4 +1,4 @@
-﻿namespace CodeMonkeys.DialogService
+﻿namespace CodeMonkeys.Dialogs
 {
     public class ConfirmationDialog : Dialog
     {
